:sparkles: ID使用Long

+++ twelvet-server/twelvet-server-system/src/main/java/com/twelvet/server/system/service/impl/SysRoleServiceImpl.java
@@ -89,7 +89,7 @@ public class SysRoleServiceImpl implements ISysRoleService {
 	 * @return 选中角色ID列表
 	 */
 	@Override
-	public List<Integer> selectRoleListByUserId(Long userId) {
+	public List<Long> selectRoleListByUserId(Long userId) {
 		return sysRoleMapper.selectRoleListByUserId(userId);
 	}
